整数へ変更

+++ database/migrations/2025_01_12_042331_create_data_table.php
@@ -15,9 +15,9 @@ return new class extends Migration
     {
         Schema::create('data', function (Blueprint $table) {
             $table->id();
-            $table->float('temperature')->nullable(false);
-            $table->float('humidity')->nullable(false);
-            $table->float('co2')->nullable(false);
+            $table->integer('temperature')->nullable(false);
+            $table->integer('humidity')->nullable(false);
+            $table->integer('co2')->nullable(false);
             $table->integer('place_id')->nullable(false);
             $table->timestamps();
         });
